python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
aiohttp==3.9.0
beautifulsoup4==4.12.2
selenium==4.15.0
openai==1.3.5
anthropic==0.7.7
python-dotenv==1.0.0
supabase==2.0.2
PyPDF2==3.0.1
python-docx==1.1.0